fix test

+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/file/object_storage/ProtoToAvroFormatter.kt
@@ -233,7 +233,7 @@ class ProtoToAvroFormatter(
                         }
                     is DateType -> { rec, p, ch ->
                             p.getDate(acc)?.let {
-                                runCatching { LocalDate.parse(it).toEpochDay() }
+                                runCatching { LocalDate.parse(it).toEpochDay().toInt() }
                                     .onSuccess { d -> rec.put(fieldName, d) }
                                     .onFailure {
                                         rec.put(fieldName, null)
